Add Unity 2022 support, disable scene view rendering and fix allocations
